add full rocksdb::WriteBatch support
The batch module in rocksdb is very useful to handle atomic write and build a
transaction system over it. Since everything is in memory and available
as a resource it can be easily be shared between process (but the data will
still not copied like in ETS and not shared). It then skips the need to create an ETS table
for such purpose (like mnesia does).

Also compared to `erocksdb:write/3`, data are only copied once and can be
garbage collected sooner.